* commit 'd51ca1556753fed5f20cd09a8608d0ee1553e41c':
  Fix incorrect parameter types for locale funcs.

+++ libc/bionic/strtoull_l.cpp
@@ -28,7 +28,7 @@
 
 #include <stdlib.h>
 
-unsigned long long strtoull_l(const char *nptr, char **endptr, size_t base,
+unsigned long long strtoull_l(const char *nptr, char **endptr, int base,
                               locale_t) {
   return strtoull(nptr, endptr, base);
 }
